See the debian changelog file
git-svn-id: https://stealth.svn.sourceforge.net/svnroot/stealth/trunk@50 93b9aca8-391c-0410-bfd7-9a5413174b8b

+++ stealth/scanner/killchildren.cc
@@ -2,6 +2,8 @@
 
 void Scanner::killChildren()
 {
-    kill(s_shPid, SIGTERM) && kill(s_shPid, SIGKILL);
-    kill(s_sshPid, SIGTERM) && kill(s_sshPid, SIGKILL);
+    if (s_shPid)
+        kill(s_shPid, SIGTERM) && kill(s_shPid, SIGKILL);
+    if (s_sshPid)
+        kill(s_sshPid, SIGTERM) && kill(s_sshPid, SIGKILL);
 }
